move lots of stuffs to expr.h/expr.cpp
